Adding a new AWS connector, and splitting the ingest types for the Boto3 connector
